requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
pendulum==2.1.2
python-dateutil==2.8.2
pytzdata==2020.1
requests==2.31.0
six==1.16.0
slack-sdk==3.22.0
urllib3==2.0.5
pillow==10.2.0
